requirements.txt
absl-py==0.7.0
aioredis==1.2.0
asgiref==2.3.2
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
baidu-aip==2.2.13.0
certifi==2016.2.28
cffi==1.12.3
channels==2.1.6
channels-redis==2.3.3
click==6.7
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
daphne==2.2.4
DBUtils==1.3
Django==2.1.7
django-cors-headers==2.4.0
django-crontab==0.7.1
django-extensions==2.1.9
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
django-werkzeug-debugger-runserver==0.3.1
docopt==0.6.2
filterpy==1.4.5
Flask==0.12.2
Flask-Cors==3.0.3
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
hbmqtt==0.9.5
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.9.6
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
labelme==3.15.2
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.0.2
more-itertools==7.0.0
msgpack==0.6.0
mysqlclient==1.4.1
numpy==1.16.0
olefile==0.44
opencv-python==4.1.0.25
paho-mqtt==1.4.0
pandas==0.20.3
passlib==1.7.1
Pillow==4.2.1
pluggy==0.9.0
protobuf==3.6.0
py==1.8.0
pycparser==2.19
PyHamcrest==1.9.0
pykalman==0.9.5
PyMySQL==0.7.9
pyOpenSSL==19.0.0
pyparsing==2.2.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
pyserial==2.7
pyStrich==0.8
pytest==4.4.1
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.13
QtPy==1.7.1
reportlab==3.4.0
requests==2.14.2
scipy==0.19.1
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
transitions==0.6.9
Twisted==18.9.0
txaio==18.8.1
websockets==7.0
Werkzeug==0.14.1
wfastcgi==3.0.0
wincertstore==0.2
WMI==1.4.9
XlsxWriter==0.9.8
xlwt==1.3.0
zope.interface==4.6.0
